Inserción de ReporteDiario para el Usuario tutor.
Create-drop, inserción al inicializar spring. Controlador de consulta de reportes por infante según tutor autenticado.

+++ src/main/java/com/virtual/virtualbaby/user/repository/CapitalHumanoRepository.java
@@ -3,6 +3,6 @@ package com.virtual.virtualbaby.user.repository;
 import com.virtual.virtualbaby.user.model.CapitalHumano;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CapitalHumanoRepository extends JpaRepository<CapitalHumano,Integer> {
+public interface CapitalHumanoRepository extends JpaRepository<CapitalHumano, Long> {
 
 }
