+++ src/main/java/com/virtual/virtualbaby/user/repository/DocenteRepository.java
@@ -3,5 +3,5 @@ package com.virtual.virtualbaby.user.repository;
 import com.virtual.virtualbaby.user.model.Docente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocenteRepository extends JpaRepository<Docente,Integer> {
+public interface DocenteRepository extends JpaRepository<Docente, Long> {
 }
