+++ src/main/java/com/virtual/virtualbaby/user/repository/TrabajadorSocialRepository.java
@@ -3,7 +3,7 @@ package com.virtual.virtualbaby.user.repository;
 import com.virtual.virtualbaby.user.model.TrabajadorSocial;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrabajadorSocialRepository extends JpaRepository<TrabajadorSocial,Integer> {
+public interface TrabajadorSocialRepository extends JpaRepository<TrabajadorSocial, Long> {
 
 }
