+++ src/main/java/com/virtual/virtualbaby/user/repository/TutorRepository.java
@@ -3,6 +3,6 @@ package com.virtual.virtualbaby.user.repository;
 import com.virtual.virtualbaby.user.model.Tutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TutorRepository extends JpaRepository<Tutor,Integer> {
+public interface TutorRepository extends JpaRepository<Tutor, Long> {
 }
